MacOS: Remove old folders only on cleanup action (#452)
PURPOSE:
2019 Maya is missing in Mac installer bundles

EFFECT OF CHANGES:
All version of Maya are present

+++ FireRender.Maya.OSX/frMayaPluginMac/prebuild.sh
@@ -12,6 +12,12 @@
 # limitations under the License.
 
 # clear old folders
+
+if [ "$ACTION" != "clean" ]; then
+	exit
+fi
+
+
 pushd "/Users/Shared/RadeonProRender"
 rm -rf cache icons image lib plug-ins renderDesc scripts shaders shelves modules
 popd
